+++ packages/server/graphql/public/types/_xLinearProject.ts
@@ -1,0 +1,7 @@
+import {_XLinearProjectResolvers} from '../resolverTypes'
+
+const _xLinearProject: _XLinearProjectResolvers = {
+  service: () => 'linear'
+}
+
+export default _xLinearProject