+++ scripts/webpack/prod.servers.config.js
@@ -15,6 +15,7 @@ const EMBEDDER_ROOT = path.join(PROJECT_ROOT, 'packages', 'embedder')
 const DOTENV = path.join(PROJECT_ROOT, 'scripts/webpack/utils/dotenv.js')
 const distPath = path.join(PROJECT_ROOT, 'dist')
 const INIT_PUBLIC_PATH = path.join(SERVER_ROOT, 'initPublicPath.ts')
+const INIT_LOGGING = path.join(SERVER_ROOT, 'logging.ts')
 
 const COMMIT_HASH = cp.execSync('git rev-parse HEAD').toString().trim()
 const runtimePlatform = `${process.platform}-${process.arch}`
@@ -31,6 +32,7 @@ module.exports = (config) => {
       web: [
         DOTENV,
         INIT_PUBLIC_PATH,
+        INIT_LOGGING,
         // each instance of web needs to generate its own index.html to use on startup
         path.join(PROJECT_ROOT, 'scripts/toolboxSrc/applyEnvVarsToClientAssets.ts'),
         path.join(SERVER_ROOT, 'server.ts')